feat: integrate AnalyzedResult and TableContainer components into App layout
